Move to utils folder
